+++ tests/test_nwchem.py
@@ -35,10 +35,9 @@ def test_opt_calc():
     assert -40.4165 < calc.get_energy() < -40.4164
     assert calc.output.exists
     assert calc.output.file_lines is not None
-    assert calc.get_imaginary_freqs() == []
     assert calc.input.filename == 'opt_nwchem.nw'
     assert calc.output.filename == 'opt_nwchem.out'
-    assert calc.terminated_normally()
+    assert calc.terminated_normally
     assert calc.optimisation_converged()
     assert calc.optimisation_nearly_converged() is False